move the thing
